Sequencer: Use double for ticks_t
This helps dealing with fractional errors resulting in dropped notes when running with
very small buffer sizes.

+++ sequencer/src/common.h
@@ -21,7 +21,7 @@
 // #pragma once
 
 typedef unsigned char byte_t;
-typedef unsigned long tick_t;
+typedef double tick_t;
 typedef unsigned int uint;
